backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Cache & Queue
redis==5.0.1
celery==5.3.6
rabbitmq==0.2.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0

# Validation & Settings
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# File Handling
minio==7.2.0
openpyxl==3.1.2
pandas==2.1.4
python-docx==1.1.0
PyPDF2==3.0.1

# AI/LLM
langchain==0.1.0
langchain-openai==0.0.2
openai==1.6.1
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1
tenacity==8.2.3
orjson==3.9.10

# Logging
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
factory-boy==3.3.0
